Change datatypes to be consistent and add explicit casts to avoid compiler warnings

+++ himan-lib/include/timer.h
@@ -35,10 +35,10 @@ public:
      * @return Elapsed time in microseconds
      */
 
-    inline long GetTime()
+    inline size_t GetTime()
     {
-        long start = static_cast<long> (start_ts.tv_sec*1000000000) + start_ts.tv_nsec;
-        long stop =  static_cast<long> (stop_ts.tv_sec*1000000000) + stop_ts.tv_nsec;
+        size_t start = static_cast<size_t> (start_ts.tv_sec*1000000000 + start_ts.tv_nsec);
+        size_t stop =  static_cast<size_t> (stop_ts.tv_sec*1000000000 + stop_ts.tv_nsec);
 
         return (stop - start) / 1000; // us
     }
